requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
idna==3.4
joblib==1.3.1
numpy==1.25.1
pandas==2.0.3
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scikit-learn==1.2.0
scipy==1.11.1
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.1.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
